+++ frontend/src/app/types/Product.ts
@@ -1,0 +1,12 @@
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  description: string;
+  quantity: number;
+  imgPath: string;
+}
+
+export interface ProductReponseDto {
+  products: Product[];
+}
